python-user-command-bug - last update broke fetching all users

+++ opal-python-client/src/main/python/opal/user.py
@@ -11,7 +11,7 @@ def add_arguments(parser):
     """
     Add data command specific options
     """
-    parser.add_argument('--name', '-n', required=True, help='User name.')
+    parser.add_argument('--name', '-n', required=False, help='User name.')
     parser.add_argument('--upassword', '-upa', required=False, help='User password.')
     parser.add_argument('--ucertificate', '-usc', required=False, help='User certificate (public key) file')
     parser.add_argument('--disabled', '-di', action='store_true', required=False,
@@ -31,7 +31,7 @@ def do_ws(args):
     """
     Build the web service resource path
     """
-    if args.add:
+    if args.add or (args.fetch and not args.name):
         ws = "/system/subject-credentials"
     else:
         ws = "/system/subject-credential/" + args.name
@@ -59,13 +59,17 @@ def do_command(args):
             # send request
             response = request.get().resource(do_ws(args)).send()
         elif args.add:
-            if not args.upassword and not args.ucertificate:
+            if not args.name:
+                raise Exception('A user name is required.')
+            if not args.upassword:
                     raise Exception('A user password or a certificat file is required.')
 
             # create user
             user = opal.protobuf.Opal_pb2.SubjectCredentialsDto()
             user.name = args.name
             if args.upassword:
+                if len(args.upassword) < 6:
+                    raise Exception('Password must contain at least 6 characters.')
                 user.authenticationType = get_authentication_type('PASSWORD')
                 user.password = args.upassword
             else:
@@ -82,6 +86,9 @@ def do_command(args):
             request.fail_on_error().accept_json().content_type_protobuf()
             response = request.post().resource(do_ws(args)).content(user.SerializeToString()).send()
         elif args.update:
+            if not args.name:
+                raise Exception('A user name is required.')
+
             userInfo = request.get().resource(do_ws(args)).send().as_json()
             request = opal.core.OpalClient.build(opal.core.OpalClient.LoginInfo.parse(args)).new_request()
             request.fail_on_error()
@@ -92,7 +99,8 @@ def do_command(args):
             if args.upassword:
                 if userInfo['authenticationType'] == "CERTIFICATE":
                     raise Exception("%s requires a certificate (public key) file" % user.name)
-
+                if len(args.upassword) < 6:
+                    raise Exception('Password must contain at least 6 characters.')
                 user.authenticationType = get_authentication_type('PASSWORD')
                 user.password = args.upassword
             else:
@@ -111,6 +119,9 @@ def do_command(args):
             request.fail_on_error().accept_json().content_type_protobuf()
             response = request.put().resource(do_ws(args)).content(user.SerializeToString()).send()
         elif args.delete:
+            if not args.name:
+                raise Exception('A user name is required.')
+
             response = request.delete().resource(do_ws(args)).send()
 
         # format response
